fix bug where last note after a digit is thrown away
Let's take a look at the input `L4C`. After `C` has been read,  since it is not a digit so the `while` loop will break. Meanwhile `scanner.eof()` is true, so `C` will not be put back to the scanner. As a result `C` is thrown away and will not be played.